chore: Flux_Limit and Popen
replacement of the run function for Popen to call the Scripts in Listening. Added also the calculation of the average flux limit.
Also replacement of the main mail, which is now sent when all scripts called by popen are finished, i.e, now the main email is written in CL_Coverage calling the function SendTheMail in the modulus with the same name